code cleaning and few bugs fixes

+++ src/com/mkr/notesdatabase/NotesDBHelper.java
@@ -11,12 +11,14 @@ import android.database.sqlite.SQLiteDatabase;
 import android.util.Log;
 
 import com.mkr.notes.Note;
+import com.mkr.notes.NotesActivity;
 
 public class NotesDBHelper {
 
 	private static NotesDBHelper mNotesDBHelper;
 	private static SQLiteHelper mSQLiteHelper;
 	private SQLiteDatabase mSQLWritableDatabase;
+	
 	private final String[] allColumns = {
 				SQLiteHelper.COLUMN_CREATED_TIME, 
 				SQLiteHelper.COLUMN_MODIFIED_TIME,
@@ -27,7 +29,7 @@ public class NotesDBHelper {
 			};
 
 	/**
-	 * this is for temp use remove it later
+	 * load all the current notes
 	 */
 	private static final Map<Long, Note> mNotesInfo = new HashMap<Long, Note>();
 	
@@ -46,6 +48,16 @@ public class NotesDBHelper {
 		mSQLWritableDatabase = mSQLiteHelper.getWritableDatabase();
 	}
 
+	/**
+	 * insert a new note into the data base
+	 * 
+	 * @param createTime created time for this note
+	 * @param modifiedTime modified time for this note 
+	 * @param noteTitle title of the note
+	 * @param notePath path for the note in the internal storage
+	 * @param label label for this note
+	 * @param isTitleSelected whether "set title" option is selected
+	 */
 	public void insertNewNote(final long createTime, final long modifiedTime, final String noteTitle,
 			final String notePath, final String label, final int isTitleSelected) {
 
@@ -60,32 +72,44 @@ public class NotesDBHelper {
 		long rowID = -1;
 		if(mSQLWritableDatabase.isOpen()) {
 			if(!isRowAlreadyExists(createTime)) {
+				//this is a new row so insert the created time as well, since this is the primary key
 				values.put(SQLiteHelper.COLUMN_CREATED_TIME, createTime);
 				rowID = mSQLWritableDatabase.insert(SQLiteHelper.TABLE_NAME, null, values);
 			} else {
-				Log.e("mkr","row already exists. So just update the values");
+				//row already exists, so just update the row. since created time is the primary key
+				//don't insert it again
 				rowID = mSQLWritableDatabase.update(SQLiteHelper.TABLE_NAME, values, SQLiteHelper.COLUMN_CREATED_TIME 
 						+ " = " + createTime, null);
 			}
 		}
 		
+		//failed to insert
 		if(rowID == -1) {
-			Log.e("mkr","FAILED TO INSERT THE NOTE");
+			Log.e(NotesActivity.TAG, "FAILED TO INSERT THE NOTE");
 		}
 	}
 
 	public void deleteNote(final long createdTime) {
-		int deletedRows = mSQLWritableDatabase.delete(SQLiteHelper.TABLE_NAME, SQLiteHelper.COLUMN_CREATED_TIME 
+		if(mSQLWritableDatabase == null) return; 
+			
+		mSQLWritableDatabase.delete(SQLiteHelper.TABLE_NAME, SQLiteHelper.COLUMN_CREATED_TIME 
 								+ " = " + createdTime, null);
-		Log.e("mkr","Number of rows deleted = " + deletedRows);
 		getAllSavedNotes();
 	}
 	
 	private boolean isRowAlreadyExists(final long createdTime) {
+		if(mNotesInfo == null) return false; 
+			
 		return mNotesInfo.containsKey(createdTime);
 	}
 	
+	/**
+	 * get all the saved notes
+	 * @return
+	 */
 	public ArrayList<Note> getAllSavedNotes() {
+		if(mSQLWritableDatabase == null) return null; 
+		
 		final ArrayList<Note> notesList = new ArrayList<Note>();
 		final Cursor cursor = mSQLWritableDatabase.query(SQLiteHelper.TABLE_NAME, allColumns, 
 				null, null, null, null, null);
@@ -104,12 +128,6 @@ public class NotesDBHelper {
 				mNotesInfo.put(note.createDate, note);
 			}
 		}
-
-		/*Log.e("mkr","Total notes count ---->"+notesList.size());
-		for (Note note : notesList) {
-			Log.e("mkr","Title ---->"+note.title);
-		}*/
-		
 		return notesList;
 	}
 
@@ -117,7 +135,16 @@ public class NotesDBHelper {
 		return mNotesInfo;
 	}
 
+	/**
+	 * get the number of notes that are associated with each labels ie.. number of work labels, personal labels etc..<br>
+	 * work label -- 2 notes<br>
+	 * ideas label - 1 note <br>
+	 * 
+	 * @return the map labels and the number of notes associated with them
+	 */
 	public Map<String, Integer> getLabelsDetails() {
+		if(mSQLWritableDatabase == null) return null; 
+		
 		final Cursor cursor = mSQLWritableDatabase.query(SQLiteHelper.TABLE_NAME, allColumns, 
 				null, null, null, null, null);
 		final Map<String, Integer> labelsMap = new HashMap<String, Integer>();
@@ -132,18 +159,12 @@ public class NotesDBHelper {
 				labelsMap.put(noteLabel, counter);
 			}
 		}
-		
-		/*Log.e("mkr","labels info");
-		Iterator<String> keysIterator = labelsMap.keySet().iterator();
-		while (keysIterator.hasNext()) {
-			String next = keysIterator.next();
-			Log.e("mkr","for " + next + " lables are --> " + labelsMap.get(next));
-		}*/
-		
 		return labelsMap;
 	}
 	
 	public void close() {
-		mSQLWritableDatabase.close();
+		if(mSQLWritableDatabase != null) {
+			mSQLWritableDatabase.close();
+		}
 	}
 }
